refactor(ooda_actions/goal_session): split source into mod + advance + gh
goal_session.rs (862) → goal_session/{mod.rs (380), advance.rs (345), gh.rs (149)}.

mod.rs keeps parse/validate/sanitize helpers + GoalSessionResult type.
advance.rs holds advance_goal_with_session + assess_only_outcome.
gh.rs holds the gh CLI dispatch helpers + is_plausible_label.

All 75 goal_session tests pass.

Refs #1266
